ADD: support floating point paramter for Plane::gaussian_blur() FIX: Planed::weighted_sum() passed the from planes line_width, causing incorrect results if image had been cropped

+++ src/RenderOperations.hpp
@@ -71,7 +71,7 @@ class RenderPipeBlurring : public ARenderPipe{
 		virtual ImageEx render( const ImageEx& img ) const override{
 			switch( method ){
 				case 1: return img.copyApplyAll( true, &Plane::blur_box, width, height ); break;
-				case 2: return img.copyApplyAll( true, &Plane::blur_gaussian, width, height ); break;
+				case 2: return img.copyApplyAll( true, &Plane::blur_gaussian, (double)width, (double)height ); break;
 				default: return img;
 			}
 		}
